cli-module: First split of the project
Everything that was under ``compressor.__init__``, now is divided into:

* compressor.cli: the command line commands
* compressor.lib: the library API functionality

Add ``compressor.__version__``, and update imports accordingly.

Breaks backwards compatibility (imports will have to be updated!).
